renamed files in homework3
